pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "crop-advisor-env"
version = "0.1.0"
description = "An RL environment for precision agriculture using Meta's OpenEnv framework"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
authors = [
    {name = "Srikanth"},
]
keywords = ["reinforcement-learning", "openenv", "agriculture", "rl-environment", "pytorch"]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["crop_advisor_env*"]
